Asset restore impl and used where needed, fixed ex18, new class CDerivativeMapCreator
