+++ src/nbl/CMakeLists.txt
@@ -182,6 +182,7 @@ set(NBL_ASSET_SOURCES
 # Images
 	${NBL_ROOT_PATH}/src/nbl/asset/interchange/IImageAssetHandlerBase.cpp
 	${NBL_ROOT_PATH}/src/nbl/asset/filters/CBasicImageFilterCommon.cpp
+	${NBL_ROOT_PATH}/src/nbl/asset/utils/CDerivativeMapCreator.cpp
 
 # Image loaders
 	${NBL_ROOT_PATH}/src/nbl/asset/interchange/IImageLoader.cpp
