+++ src/nbl/asset/interchange/CGraphicsPipelineLoaderMTL.cpp
@@ -12,6 +12,7 @@
 #include "nbl/asset/asset.h"
 #include "nbl/asset/interchange/CGraphicsPipelineLoaderMTL.h"
 #include "nbl/asset/utils/IGLSLEmbeddedIncludeLoader.h"
+#include "nbl/asset/utils/CDerivativeMapCreator.h"
 
 #include "nbl/builtin/MTLdefaults.h"
 
@@ -521,7 +522,15 @@ CGraphicsPipelineLoaderMTL::image_views_set_t CGraphicsPipelineLoaderMTL::loadIm
         SAssetLoadParams lp;
         if (_mtl.maps[i].size() )
         {
-            auto bundle = interm_getAssetInHierarchy(m_assetMgr, relDir+_mtl.maps[i], lp, _ctx.topHierarchyLevel+ICPURenderpassIndependentPipeline::IMAGE_HIERARCHYLEVELS_BELOW, _ctx.loaderOverride);
+            SAssetBundle bundle;
+            if (i != CMTLMetadata::CRenderpassIndependentPipeline::EMP_BUMP)
+                bundle = interm_getAssetInHierarchy(m_assetMgr, relDir+_mtl.maps[i], lp, _ctx.topHierarchyLevel+ICPURenderpassIndependentPipeline::IMAGE_HIERARCHYLEVELS_BELOW, _ctx.loaderOverride);
+            else
+            {
+                // we need bumpmap restored to create derivative map from it
+                const uint32_t restoreLevels = 3u; // 2 in case of image (image, texel buffer) and 3 in case of image view (view, image, texel buffer)
+                bundle = interm_getRestoredAssetInHierarchy(restoreLevels, m_assetMgr, relDir+_mtl.maps[i], lp, _ctx.topHierarchyLevel+ICPURenderpassIndependentPipeline::IMAGE_HIERARCHYLEVELS_BELOW, _ctx.loaderOverride);
+            }
             auto asset = _ctx.loaderOverride->chooseDefaultAsset(bundle,_ctx.inner);
             if (asset)
             switch (bundle.getAssetType())
@@ -617,7 +626,10 @@ CGraphicsPipelineLoaderMTL::image_views_set_t CGraphicsPipelineLoaderMTL::loadIm
 
     for (uint32_t i = 0u; i < views.size(); ++i)
     {
-        if (!images[i])
+        core::smart_refctd_ptr<ICPUImage> image = images[i];
+        if (i == CMTLMetadata::CRenderpassIndependentPipeline::EMP_BUMP && views[i])
+            image = views[i]->getCreationParameters().image;
+        if (!image)
             continue;
 
         const std::string viewCacheKey = _mtl.maps[i] + "?view";
@@ -631,6 +643,12 @@ CGraphicsPipelineLoaderMTL::image_views_set_t CGraphicsPipelineLoaderMTL::loadIm
             }
         }
 
+        if (i == CMTLMetadata::CRenderpassIndependentPipeline::EMP_BUMP)
+        {
+            const ISampler::E_TEXTURE_CLAMP wrap = _mtl.isClampToBorder(CMTLMetadata::CRenderpassIndependentPipeline::EMP_BUMP) ? ISampler::ETC_CLAMP_TO_BORDER : ISampler::ETC_REPEAT;
+            image = CDerivativeMapCreator::createDerivativeMapFromHeightMap(image.get(), wrap, wrap, ISampler::ETBC_FLOAT_OPAQUE_BLACK);
+        }
+
         constexpr IImageView<ICPUImage>::E_TYPE viewType[2]{ IImageView<ICPUImage>::ET_2D, IImageView<ICPUImage>::ET_CUBE_MAP };
         constexpr uint32_t layerCount[2]{ 1u, 6u };
 
@@ -638,13 +656,13 @@ CGraphicsPipelineLoaderMTL::image_views_set_t CGraphicsPipelineLoaderMTL::loadIm
 
         ICPUImageView::SCreationParams viewParams;
         viewParams.flags = static_cast<ICPUImageView::E_CREATE_FLAGS>(0u);
-        viewParams.format = images[i]->getCreationParameters().format;
+        viewParams.format = image->getCreationParameters().format;
         viewParams.viewType = viewType[isCubemap];
         viewParams.subresourceRange.baseArrayLayer = 0u;
         viewParams.subresourceRange.layerCount = layerCount[isCubemap];
         viewParams.subresourceRange.baseMipLevel = 0u;
         viewParams.subresourceRange.levelCount = 1u;
-        viewParams.image = std::move(images[i]);
+        viewParams.image = std::move(image);
 
         views[i] = ICPUImageView::create(std::move(viewParams));
         _ctx.loaderOverride->insertAssetIntoCache(SAssetBundle(nullptr,{ views[i] }), viewCacheKey, _ctx.inner, _ctx.topHierarchyLevel+ICPURenderpassIndependentPipeline::IMAGEVIEW_HIERARCHYLEVELS_BELOW);
