+++ src/nbl/ext/MitsubaLoader/CMitsubaLoader.cpp
@@ -6,12 +6,11 @@
 #include "os.h"
 
 #include <cwchar>
-#include <nbl/asset/filters/CSwizzleAndConvertImageFilter.h>
 
 #include "nbl/ext/MitsubaLoader/CMitsubaLoader.h"
 #include "nbl/ext/MitsubaLoader/ParserUtil.h"
 
-#include "nbl/asset/interchange/IImageAssetHandlerBase.h"
+#include "nbl/asset/utils/CDerivativeMapCreator.h"
 
 #include "nbl/ext/MitsubaLoader/CMitsubaSerializedMetadata.h"
 #include "nbl/ext/MitsubaLoader/CGLSLMitsubaLoaderBuiltinIncludeLoader.h"
@@ -250,208 +249,6 @@ static core::smart_refctd_ptr<asset::ICPURenderpassIndependentPipeline> createPi
 	return pipeline;
 }
 
-
-//TODO move all derivative-related code to different file (maybe even to engine)
-// @Crisspl doesn't OBJ also use derivative mapping??
-// @Anastazluk will nneed this for glTF
-namespace
-{
-template<class Kernel>
-class MyKernel : public asset::CFloatingPointSeparableImageFilterKernelBase<MyKernel<Kernel>>
-{
-		using Base = asset::CFloatingPointSeparableImageFilterKernelBase<MyKernel<Kernel>>;
-
-		Kernel kernel;
-		float multiplier;
-
-	public:
-		using value_type = typename Base::value_type;
-
-		MyKernel(Kernel&& k, float _imgExtent) : Base(k.negative_support.x, k.positive_support.x), kernel(std::move(k)), multiplier(_imgExtent) {}
-
-		// no special user data by default
-		inline const asset::IImageFilterKernel::UserData* getUserData() const { return nullptr; }
-
-		inline float weight(float x, int32_t channel) const
-		{
-			return kernel.weight(x, channel) * multiplier;
-		}
-		
-		// we need to ensure to override the default behaviour of `CFloatingPointSeparableImageFilterKernelBase` which applies the weight along every axis
-		template<class PreFilter, class PostFilter>
-		struct sample_functor_t
-		{
-				sample_functor_t(const MyKernel* _this, PreFilter& _preFilter, PostFilter& _postFilter) :
-					_this(_this), preFilter(_preFilter), postFilter(_postFilter) {}
-
-				inline void operator()(value_type* windowSample, core::vectorSIMDf& relativePos, const core::vectorSIMDi32& globalTexelCoord, const asset::IImageFilterKernel::UserData* userData)
-				{
-					preFilter(windowSample, relativePos, globalTexelCoord, userData);
-					auto* scale = asset::IImageFilterKernel::ScaleFactorUserData::cast(userData);
-					for (int32_t i=0; i<MaxChannels; i++)
-					{
-						// this differs from the `CFloatingPointSeparableImageFilterKernelBase`
-						windowSample[i] *= _this->weight(relativePos.x, i);
-						if (scale)
-							windowSample[i] *= scale->factor[i];
-					}
-					postFilter(windowSample, relativePos, globalTexelCoord, userData);
-				}
-
-			private:
-				const MyKernel* _this;
-				PreFilter& preFilter;
-				PostFilter& postFilter;
-		};
-
-		_NBL_STATIC_INLINE_CONSTEXPR bool has_derivative = false;
-
-		NBL_DECLARE_DEFINE_CIMAGEFILTER_KERNEL_PASS_THROUGHS(Base)
-};
-	
-template<class Kernel>
-class SeparateOutXAxisKernel : public asset::CFloatingPointSeparableImageFilterKernelBase<SeparateOutXAxisKernel<Kernel>>
-{
-		using Base = asset::CFloatingPointSeparableImageFilterKernelBase<SeparateOutXAxisKernel<Kernel>>;
-
-		Kernel kernel;
-
-	public:
-		// passthrough everything
-		using value_type = typename Kernel::value_type;
-
-		_NBL_STATIC_INLINE_CONSTEXPR auto MaxChannels = Kernel::MaxChannels; // derivative map only needs 2 channels
-
-		SeparateOutXAxisKernel(Kernel&& k) : Base(k.negative_support.x, k.positive_support.x), kernel(std::move(k)) {}
-
-		NBL_DECLARE_DEFINE_CIMAGEFILTER_KERNEL_PASS_THROUGHS(Base)
-					
-		// we need to ensure to override the default behaviour of `CFloatingPointSeparableImageFilterKernelBase` which applies the weight along every axis
-		template<class PreFilter, class PostFilter>
-		struct sample_functor_t
-		{
-				sample_functor_t(const SeparateOutXAxisKernel<Kernel>* _this, PreFilter& _preFilter, PostFilter& _postFilter) :
-					_this(_this), preFilter(_preFilter), postFilter(_postFilter) {}
-
-				inline void operator()(value_type* windowSample, core::vectorSIMDf& relativePos, const core::vectorSIMDi32& globalTexelCoord, const asset::IImageFilterKernel::UserData* userData)
-				{
-					preFilter(windowSample, relativePos, globalTexelCoord, userData);
-					auto* scale = asset::IImageFilterKernel::ScaleFactorUserData::cast(userData);
-					for (int32_t i=0; i<MaxChannels; i++)
-					{
-						// this differs from the `CFloatingPointSeparableImageFilterKernelBase`
-						windowSample[i] *= _this->kernel.weight(relativePos.x, i);
-						if (scale)
-							windowSample[i] *= scale->factor[i];
-					}
-					postFilter(windowSample, relativePos, globalTexelCoord, userData);
-				}
-
-			private:
-				const SeparateOutXAxisKernel<Kernel>* _this;
-				PreFilter& preFilter;
-				PostFilter& postFilter;
-		};
-
-		// the method all kernels must define and overload
-		template<class PreFilter, class PostFilter>
-		inline auto create_sample_functor_t(PreFilter& preFilter, PostFilter& postFilter) const
-		{
-			return sample_functor_t(this,preFilter,postFilter);
-		}
-};
-
-}
-
-static core::smart_refctd_ptr<asset::ICPUImage> createDerivMapFromHeightMap(asset::ICPUImage* _inImg, asset::ISampler::E_TEXTURE_CLAMP _uwrap, asset::ISampler::E_TEXTURE_CLAMP _vwrap, asset::ISampler::E_TEXTURE_BORDER_COLOR _borderColor)
-{
-	using namespace asset;
-
-	auto getRGformat = [](asset::E_FORMAT f) -> asset::E_FORMAT {
-		const uint32_t bytesPerChannel = (getBytesPerPixel(f) * core::rational(1, getFormatChannelCount(f))).getIntegerApprox();
-		switch (bytesPerChannel)
-		{
-		case 1u:
-			return asset::EF_R8G8_SNORM;
-		case 2u:
-			return asset::EF_R16G16_SNORM;
-		case 4u:
-			return asset::EF_R32G32_SFLOAT;
-		case 8u:
-			return asset::EF_R64G64_SFLOAT;
-		default:
-			return asset::EF_UNKNOWN;
-		}
-	};
-
-	using ReconstructionKernel = CGaussianImageFilterKernel<>; // or Mitchell
-	using DerivKernel_ = CDerivativeImageFilterKernel<ReconstructionKernel>;
-	using DerivKernel = MyKernel<DerivKernel_>;
-	using XDerivKernel_ = CChannelIndependentImageFilterKernel<DerivKernel, CBoxImageFilterKernel>;
-	using YDerivKernel_ = CChannelIndependentImageFilterKernel<CBoxImageFilterKernel, DerivKernel>;
-	using XDerivKernel = SeparateOutXAxisKernel<XDerivKernel_>;
-	using YDerivKernel = SeparateOutXAxisKernel<YDerivKernel_>;
-	constexpr bool NORMALIZE = false;
-	using DerivativeMapFilter = CBlitImageFilter
-		<
-		NORMALIZE, false, DefaultSwizzle, IdentityDither, // (Criss, look at impl::CSwizzleAndConvertImageFilterBase)
-		XDerivKernel,
-		YDerivKernel,
-		CBoxImageFilterKernel
-		>;
-
-	const auto extent = _inImg->getCreationParameters().extent;
-	const float mlt = 1.f;// static_cast<float>(std::max(extent.width, extent.height));
-	XDerivKernel xderiv(XDerivKernel_(DerivKernel(DerivKernel_(ReconstructionKernel()), mlt), CBoxImageFilterKernel()));
-	YDerivKernel yderiv(YDerivKernel_(CBoxImageFilterKernel(), DerivKernel(DerivKernel_(ReconstructionKernel()), mlt)));
-
-	using swizzle_t = asset::ICPUImageView::SComponentMapping;
-	DerivativeMapFilter::state_type state(std::move(xderiv), std::move(yderiv), CBoxImageFilterKernel());
-
-	state.swizzle = { swizzle_t::ES_R, swizzle_t::ES_R, swizzle_t::ES_R, swizzle_t::ES_R };
-
-	const auto& inParams = _inImg->getCreationParameters();
-	auto outParams = inParams;
-	outParams.format = getRGformat(outParams.format);
-	const uint32_t pitch = IImageAssetHandlerBase::calcPitchInBlocks(outParams.extent.width, asset::getTexelOrBlockBytesize(outParams.format));
-	auto buffer = core::make_smart_refctd_ptr<asset::ICPUBuffer>(asset::getTexelOrBlockBytesize(outParams.format) * pitch * outParams.extent.height);
-	asset::ICPUImage::SBufferCopy region;
-	region.imageOffset = { 0,0,0 };
-	region.imageExtent = outParams.extent;
-	region.imageSubresource.baseArrayLayer = 0u;
-	region.imageSubresource.layerCount = 1u;
-	region.imageSubresource.mipLevel = 0u;
-	region.bufferRowLength = pitch;
-	region.bufferImageHeight = 0u;
-	region.bufferOffset = 0u;
-	auto outImg = asset::ICPUImage::create(std::move(outParams));
-	outImg->setBufferAndRegions(std::move(buffer), core::make_refctd_dynamic_array<core::smart_refctd_dynamic_array<IImage::SBufferCopy>>(1ull, region));
-
-	state.inOffset = { 0,0,0 };
-	state.inBaseLayer = 0u;
-	state.outOffset = { 0,0,0 };
-	state.outBaseLayer = 0u;
-	state.inExtent = inParams.extent;
-	state.outExtent = state.inExtent;
-	state.inLayerCount = 1u;
-	state.outLayerCount = 1u;
-	state.inMipLevel = 0u;
-	state.outMipLevel = 0u;
-	state.inImage = _inImg;
-	state.outImage = outImg.get();
-	state.axisWraps[0] = _uwrap;
-	state.axisWraps[1] = _vwrap;
-	state.axisWraps[2] = asset::ISampler::ETC_CLAMP_TO_EDGE;
-	state.borderColor = _borderColor;
-	state.scratchMemoryByteSize = DerivativeMapFilter::getRequiredScratchByteSize(&state);
-	state.scratchMemory = reinterpret_cast<uint8_t*>(_NBL_ALIGNED_MALLOC(state.scratchMemoryByteSize, _NBL_SIMD_ALIGNMENT));
-
-	DerivativeMapFilter::execute(&state);
-
-	_NBL_ALIGNED_FREE(state.scratchMemory);
-
-	return outImg;
-}
 static core::smart_refctd_ptr<asset::ICPUImageView> createImageView(core::smart_refctd_ptr<asset::ICPUImage>&& _img) // TODO: this should seriously be a utility somewhere
 {
 	const auto& iparams = _img->getCreationParameters();
@@ -473,7 +270,7 @@ static core::smart_refctd_ptr<asset::ICPUImage> createDerivMap(asset::ICPUImage*
 {
 	const auto& sp = _smplr->getParams();
 
-	return createDerivMapFromHeightMap(
+	return asset::CDerivativeMapCreator::createDerivativeMapFromHeightMap(
 			_heightMap,
 			static_cast<asset::ICPUSampler::E_TEXTURE_CLAMP>(sp.TextureWrapU),
 			static_cast<asset::ICPUSampler::E_TEXTURE_CLAMP>(sp.TextureWrapV),
@@ -732,6 +529,7 @@ asset::SAssetBundle CMitsubaLoader::loadAsset(io::IReadFile* _file, const asset:
 		);
 		ctx.meta->m_global.m_materialCompilerGLSL_declarations = compResult.fragmentShaderSource_declarations;
 		ctx.meta->m_global.m_materialCompilerGLSL_source = compResult.fragmentShaderSource;
+		ctx.meta->m_global.m_ds0 = ds0;
 
 		auto meshSmartPtrArray = core::make_refctd_dynamic_array<SAssetBundle::contents_container_t>(meshes.size());
 		auto meshSmartPtrArrayIt = meshSmartPtrArray->begin();
@@ -1085,7 +883,7 @@ SContext::shape_ass_type CMitsubaLoader::loadBasicShape(SContext& ctx, uint32_t 
 	return mesh;
 }
 
-SContext::tex_ass_type CMitsubaLoader::cacheTexture(SContext& ctx, uint32_t hierarchyLevel, const CElementTexture* tex)
+SContext::tex_ass_type CMitsubaLoader::cacheTexture(SContext& ctx, uint32_t hierarchyLevel, const CElementTexture* tex, bool _restore)
 {
 	if (!tex)
 		return {};
@@ -1139,12 +937,29 @@ SContext::tex_ass_type CMitsubaLoader::cacheTexture(SContext& ctx, uint32_t hier
 					auto contents = viewBundle.getContents();
 					if (!contents.empty())
 						view = core::smart_refctd_ptr_static_cast<asset::ICPUImageView>(contents.begin()[0]);
+					if (view)
+					{
+						auto& image = view->getCreationParameters().image;
+						if (_restore && image->isADummyObjectForCache())
+						{
+							// this will restore the image being kept by found `view`
+							auto bundle = interm_getRestoredAssetInHierarchy(2u, m_assetMgr, tex->bitmap.filename.svalue, ctx.inner.params, hierarchyLevel, ctx.override_);
+							if (bundle.getContents().empty() || image->isADummyObjectForCache())
+							{
+								// if for some reason restore failed, force recreating whole view
+								auto removeBundle = asset::SAssetBundle(nullptr, { view });
+								m_assetMgr->removeAssetFromCache(removeBundle);
+								view = nullptr;
+							}
+						}
+					}
 				}
 
 				core::smart_refctd_ptr<asset::ICPUImage> img;
 				if (!view)
 				{
-					asset::SAssetBundle imgBundle = interm_getAssetInHierarchy(m_assetMgr,tex->bitmap.filename.svalue,ctx.inner.params,hierarchyLevel,ctx.override_);
+					const uint32_t restoreLevels = _restore ? 2u : 0u;
+					asset::SAssetBundle imgBundle = interm_getRestoredAssetInHierarchy(restoreLevels,m_assetMgr,tex->bitmap.filename.svalue,ctx.inner.params,hierarchyLevel,ctx.override_);
 					auto contentRange = imgBundle.getContents();
 					if (contentRange.begin() < contentRange.end())
 					{
